requirements.txt
# Web Framework
flask==3.0.2
gevent==24.2.1

# AI and Language Processing
openai>=1.58.1,<2.0.0
langchain>=0.3.18,<0.4.0
langchain-community==0.3.17
langchain-core>=0.3.0,<0.4.0
langchain-openai>=0.2.0,<0.3.0
langchain-text-splitters>=0.3.3,<0.4.0
langchain_hyperbrowser==0.3
volcengine-python-sdk[ark]==1.0.125
# Document Processing
unstructured[base]>=0.11.8,<0.17.0
openpyxl==3.1.2
PyMuPDF==1.22.3

# Database and Vector Store
pymongo==4.6.3
mongoengine==0.28

# Utilities and Dependencies
pydantic>=2.6.4,<3.0.0
urllib3==1.26.18
cnos-connector==0.1.9
python-docx
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0

# 支付和认证
stripe>=8.0.0
alipay-sdk-python>=3.6.0

# 图像处理
pillow>=10.0.0
qrcode>=7.3.1
opencv-python>=4.8.0

# Office处理
python-pptx>=0.6.21